Moved things around; tests now pass!
