+++ tests/test_main.py
@@ -1,12 +1,10 @@
 import datetime
-import pytest
 from packaging.version import Version
 
 from spec0.releasesource import PyPIReleaseSource, CondaReleaseSource, Release
 from spec0.releasefilters import SPEC0StrictDate
 
 from spec0.main import *
-from spec0.main import _major_minor_str
 
 
 class DummySource:
@@ -52,20 +50,6 @@ def test_default_filter():
     assert filter_obj.python_override is True
 
 
-@pytest.mark.parametrize(
-    "version_str,expected",
-    [
-        ("1.2", "1.2"),  # For "1.2", epoch is 0, so just "major.minor"
-        ("1!1.2", "1!1.2"),  # For "1!1.2", non-zero epoch is included
-        ("0.3.4", "0.3"),  # For "0.3.4", only major and minor are used
-    ],
-)
-def test_major_minor_str(version_str, expected):
-    v = Version(version_str)
-    result = _major_minor_str(v)
-    assert result == expected
-
-
 def test_main_uses_first_source_with_releases():
     v = Version("1.0")
     release_date = datetime.datetime(2020, 1, 1)
@@ -103,51 +87,3 @@ def test_main_uses_second_source_when_first_returns_none():
     assert result["package"] == "testpkg"
     assert isinstance(result["releases"], list)
     assert len(result["releases"]) == 1
-
-
-@pytest.mark.parametrize("release_date", [True, False])
-@pytest.mark.parametrize("drop_date", [True, False])
-def test_terminal_output_combined(capsys, release_date, drop_date):
-    pkg_info = {
-        "package": "mypackage",
-        "releases": [
-            {
-                "version": Version("1.0"),
-                "release-date": datetime.datetime(2020, 1, 1),
-                "drop-date": datetime.datetime(2021, 1, 1),
-            },
-            {
-                "version": Version("1!2.3"),
-                "release-date": datetime.datetime(2020, 2, 2),
-                "drop-date": datetime.datetime(2021, 2, 2),
-            },
-        ],
-    }
-
-    terminal_output(pkg_info, release_date=release_date, drop_date=drop_date)
-    captured = capsys.readouterr().out
-    header_line, _, first_row, second_row = captured.splitlines()
-
-    # Fixed assertions that are always expected.
-    assert "Package" in header_line
-    assert "mypackage 1.0" in first_row
-    assert "mypackage 1!2.3" in second_row
-
-    # Check header for release date and drop date based on flags.
-    if release_date:
-        assert "Release Date" in header_line
-        assert "2020-01-01" in first_row
-        assert "2020-02-02" in second_row
-    else:
-        assert "Release Date" not in header_line
-        assert "2020-01-01" not in first_row
-        assert "2020-02-02" not in second_row
-
-    if drop_date:
-        assert "Drop Date" in header_line
-        assert "2021-01-01" in first_row
-        assert "2021-02-02" in second_row
-    else:
-        assert "Drop Date" not in header_line
-        assert "2021-01-01" not in first_row
-        assert "2021-02-02" not in second_row
